Refactor code in server
