+++ server/database.py
@@ -2,7 +2,7 @@ import os
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
-DATABASE_URL = "sqlite:///./databases/test.db" if os.getenv("TESTING") == "True" else "sqlite:///./databases/main.db"
+DATABASE_URL = "sqlite:///./test.db" if os.getenv("TESTING") == "True" else "sqlite:///./main.db"
 
 engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
