+++ server/test_auth.py
@@ -3,7 +3,7 @@ from fastapi.testclient import TestClient
 from main import app
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from server.core.database import Base, get_db
+from database import Base, get_db
 import os
 
 # Set the environment to use the test database
