+++ server/test_profiles.py
@@ -4,9 +4,9 @@ from fastapi.testclient import TestClient
 from main import app
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, Session
-from server.core.database import Base, get_db
-from server.core.models import User
-from server.utils.hashing import hash_password
+from database import Base, get_db
+from models import User
+from hashing import hash_password
 import random
 import string
 import uuid
